made the settings a drawer and made the "pixel" area a component
