--- src/lib/components/SettingsModal.svelte
@@ -1,165 +1,0 @@
-<script>
-	import Icon from '@iconify/svelte';
-
-	let {
-		openModal = $bindable(),
-		numIconsPerRow = $bindable(),
-		speed = $bindable(),
-		frequency = $bindable(),
-		dotOptions = $bindable({
-			color: 'ff0000',
-			icon: 'material-symbols:square-rounded'
-		}),
-		dashOptions = $bindable({
-			color: '0000ff',
-			icon: 'material-symbols:square-rounded'
-		}),
-		spaceOptions = $bindable({
-			color: '000000',
-			icon: 'material-symbols:square-rounded'
-		}),
-		backgroundColor = $bindable('000000')
-	} = $props();
-
-	let dialog = $state();
-
-	function closeModal() {
-		dialog.close();
-	}
-
-	function handleBackdropClick(e) {
-		if (e.target === dialog) dialog.close();
-	}
-
-	function handleHexInput(value) {
-		return value.slice(1);
-	}
-
-	$effect(() => {
-		if (openModal) dialog.showModal();
-	});
-</script>
-
-<dialog
-	bind:this={dialog}
-	onclose={() => (openModal = false)}
-	onclick={handleBackdropClick}
-	class="mx-auto my-auto w-2/3 rounded-lg border-2 border-black bg-gray-400 p-2 text-shadow-none"
->
-	<div class="space-y-2">
-		<div class="flex items-center justify-between">
-			<h1 class="text-2xl font-semibold">Settings</h1>
-			<button onclick={closeModal} aria-label="Close settings" class="cursor-pointer">
-				<Icon icon="material-symbols:close-rounded" width="30" />
-			</button>
-		</div>
-		<hr />
-		<div class="justify-evenly font-semibold lg:flex">
-			{#each [{ title: 'Dot / 0', options: dotOptions }, { title: 'Dash / 1', options: dashOptions }, { title: 'Space', options: spaceOptions }] as section (section.title)}
-				<div>
-					<h2 class="text-xl font-bold">{section.title}</h2>
-					<div class="mr-10 ml-10 flex items-center justify-around">
-						<div class="space-y-2">
-							<div class="flex items-center justify-between">
-								<h3>Icon:</h3>
-								<input type="text" bind:value={section.options.icon} class="w-3/4 rounded-lg" placeholder="Icon name" />
-							</div>
-							<div class="flex items-center justify-between">
-								<h3>Color:</h3>
-								<input
-									type="color"
-									value={`#${section.options.color}`}
-									onchange={(e) => (section.options.color = handleHexInput(e.target.value))}
-									class="rounded-lg"
-								/>
-							</div>
-							<div class="flex items-center justify-between">
-								<h3>Show:</h3>
-								<input type="checkbox" bind:checked={section.options.show} />
-							</div>
-						</div>
-						<div class="mr-auto ml-10 hidden w-1/4 md:block">
-							<Icon icon={section.options.icon} style="color: #{section.options.color}" width="fit" />
-						</div>
-					</div>
-				</div>
-			{/each}
-		</div>
-		<hr />
-		<div class="space-y-3 font-semibold">
-			<div class="flex items-center gap-4">
-				<h2 class="min-w-fit text-xl font-bold">Background Color:</h2>
-				<input
-					type="color"
-					value={`#${backgroundColor}`}
-					onchange={(e) => (backgroundColor = handleHexInput(e.target.value))}
-					class="rounded-lg"
-				/>
-			</div>
-			<div class="space-y-2">
-				<div class="flex items-center gap-4">
-					<h2 class="min-w-fit text-xl font-bold">Frequency:</h2>
-					<input type="number" bind:value={frequency} class="w-20 rounded-lg" />
-				</div>
-				<div class="flex items-center gap-2">
-					<span>100</span>
-					<input type="range" min="100" max="800" step="1" bind:value={frequency} class="flex-1" />
-					<span>800</span>
-				</div>
-			</div>
-			<div class="space-y-2">
-				<div class="flex items-center gap-4">
-					<h2 class="min-w-fit text-xl font-bold">Icons Per Row:</h2>
-					<input type="number" bind:value={numIconsPerRow} class="w-20 rounded-lg" />
-				</div>
-				<div class="flex items-center gap-2">
-					<span>1</span>
-					<input type="range" min="1" max="50" step="1" bind:value={numIconsPerRow} class="flex-1" />
-					<span>50</span>
-				</div>
-			</div>
-			<div class="space-y-2">
-				<div class="flex items-center gap-4">
-					<h2 class="min-w-fit text-xl font-semibold">Speed:</h2>
-					<input type="number" bind:value={speed} class="w-20 rounded-lg" />
-				</div>
-				<div class="flex items-center gap-2">
-					<span>0</span>
-					<input type="range" min="0" max="1" step="0.01" bind:value={speed} class="flex-1" />
-					<span>1</span>
-				</div>
-			</div>
-		</div>
-	</div>
-</dialog>
-
-<style>
-	dialog::backdrop {
-		background: rgba(0, 0, 0, 0.3);
-	}
-	dialog[open] {
-		animation: zoom 0.3s cubic-bezier(0.34, 1.56, 0.64, 1);
-	}
-	@keyframes zoom {
-		from {
-			transform: scale(0.95);
-		}
-		to {
-			transform: scale(1);
-		}
-	}
-	dialog[open]::backdrop {
-		animation: fade 0.2s ease-out;
-	}
-	@keyframes fade {
-		from {
-			opacity: 0;
-		}
-		to {
-			opacity: 1;
-		}
-	}
-	button {
-		display: block;
-	}
-</style>
