+++ src/routes/+page.svelte
@@ -1,8 +1,8 @@
 <script>
-	import SettingsModal from '$lib/components/SettingsModal.svelte';
 	import Icon from '@iconify/svelte';
-	import html2canvas from 'html2canvas-pro';
 	import * as Tone from 'tone';
+	import FrameOutput from '$lib/components/FrameOutput.svelte';
+	import SettingsDrawer from '$lib/components/SettingsDrawer.svelte';
 
 	const morseCodeValues = $state({
 		A: '.-',
@@ -45,7 +45,6 @@
 	});
 	const options = $state(['Text', 'Morse', 'Binary']);
 
-	let captureTarget = $state(); // Use HTMLElement for type safety
 	let userInput = $state('');
 	let userOutput = $state('');
 	let inputOption = $state('Text');
@@ -57,7 +56,7 @@
 	let speed = $state(0.25);
 	let isPlaying = $state(false);
 	let playingIndex = $state(-1);
-	let openModal = $state(false);
+	let openDrawer = $state(false);
 	let frequency = $state(600);
 	let dotOptions = $state({
 		color: 'ff0000',
@@ -92,6 +91,13 @@
 			.join(' ');
 	}
 
+	function morseToText() {
+		textString = morseString
+			.split(' ')
+			.map((letter) => Object.keys(morseCodeValues).find((key) => morseCodeValues[key] === letter) || '')
+			.join('');
+	}
+
 	function convertInput() {
 		if (inputOption == 'Text') {
 			userInput = userInput.toUpperCase().replaceAll(/[^A-Z 0-9]+/g, '');
@@ -101,10 +107,7 @@
 		} else if (inputOption == 'Morse') {
 			userInput = userInput.toUpperCase().replaceAll(/[^\.\- ]+/g, '');
 			morseString = userInput;
-			textString = userInput
-				.split(' ')
-				.map((letter) => Object.keys(morseCodeValues).find((key) => morseCodeValues[key] === letter) || '')
-				.join('');
+			morseToText();
 			textToBinary();
 		} else if (inputOption == 'Binary') {
 			userInput = userInput.replaceAll(/[^0-1 ]+/g, '');
@@ -150,32 +153,10 @@
 			isPlaying = false;
 		}
 	}
-
-	async function takeScreenshot() {
-		if (!captureTarget) {
-			console.error('Target element not found!');
-			return;
-		}
-
-		// Use html2canvas with options to handle potential issues like CORS for images
-		const canvas = await html2canvas(captureTarget, {
-			useCORS: true, // Attempt to load images using CORS
-			allowTaint: true // Allow cross-origin images to "taint" the canvas
-		});
-
-		const image = canvas.toDataURL('image/png');
-		const link = document.createElement('a');
-		link.download = 'MorseCodeImage.png';
-		link.href = image;
-		link.click(); // Programmatically click the link to trigger download
-
-		// Clean up the temporary link element
-		link.remove();
-	}
 </script>
 
-<SettingsModal
-	bind:openModal
+<SettingsDrawer
+	bind:openDrawer
 	bind:dotOptions
 	bind:dashOptions
 	bind:spaceOptions
@@ -187,7 +168,7 @@
 
 <div class="space-y-5">
 	<div class="flex text-gray-400">
-		<button class="ml-auto cursor-pointer" aria-label="Settings" onclick={() => (openModal = true)}>
+		<button class="ml-auto cursor-pointer" aria-label="Settings" onclick={() => (openDrawer = true)}>
 			<Icon icon="mdi:gear" width="40" />
 		</button>
 	</div>
@@ -207,7 +188,9 @@
 				</select>
 			</div>
 			<textarea
-				class="min-h-[200px] w-full max-w-3xl rounded-lg border-2 border-black text-lg uppercase shadow-lg/100"
+				class="min-h-[200px] w-full max-w-3xl rounded-lg border-2 border-black text-lg {inputOption == 'Morse'
+					? 'font-bold'
+					: 'font-semibold'} uppercase shadow-lg/100"
 				bind:value={userInput}
 				oninput={convertInput}
 				aria-label="Input text"
@@ -247,7 +230,9 @@
 			</div>
 			<textarea
 				disabled
-				class="min-h-[200px] w-full max-w-3xl rounded-lg border-2 border-black bg-gray-300 text-lg text-black uppercase shadow-lg/100"
+				class="min-h-[200px] w-full max-w-3xl rounded-lg border-2 border-black bg-gray-400 text-lg {outputOption == 'Morse'
+					? 'font-bold'
+					: 'font-semibold'} text-black uppercase shadow-lg/100"
 				bind:value={userOutput}
 				aria-label="Output text"
 			></textarea>
@@ -269,38 +254,14 @@
 			</div>
 		</div>
 	</div>
-	{#if userOutput.length > 0}
-		<button
-			class="mx-auto flex w-fit items-center justify-center space-x-2 rounded-lg p-2 text-center text-black {isPlaying ||
-			userOutput.length == 0
-				? 'bg-green-800'
-				: 'cursor-pointer bg-green-500 hover:bg-green-300 '}"
-			onclick={takeScreenshot}
-			disabled={isPlaying || userOutput.length == 0}
-			><Icon icon="lsicon:picture-filled" width="20" />
-			<p>Save Image</p></button
-		>
-		<div class="flex justify-center pb-5">
-			<div
-				bind:this={captureTarget}
-				style="display: grid; grid-template-columns: repeat({numIconsPerRow}, 1fr); background-color: #{backgroundColor};"
-			>
-				<!-- Pixel Area -->
-				{#each userOutput as symbol, index}
-					{#if (symbol == '.' || symbol == '0') && dotOptions.show}
-						<Icon icon={dotOptions.icon} class={index == playingIndex ? 'scale-125' : ''} width="fit" style="color: #{dotOptions.color}" />
-					{:else if (symbol == '-' || symbol == '1') && dashOptions.show}
-						<Icon
-							icon={dashOptions.icon}
-							class={index == playingIndex ? 'scale-125' : ''}
-							width="fit"
-							style="color: #{dashOptions.color}"
-						/>
-					{:else if symbol == ' ' && spaceOptions.show}
-						<Icon icon={spaceOptions.icon} style="color: #{spaceOptions.color}" width="fit" />
-					{/if}
-				{/each}
-			</div>
-		</div>
-	{/if}
+	<FrameOutput
+		bind:userOutput
+		bind:isPlaying
+		bind:playingIndex
+		bind:dotOptions
+		bind:dashOptions
+		bind:spaceOptions
+		bind:numIconsPerRow
+		bind:backgroundColor
+	/>
 </div>
